fix: enforce architectural rules - remove logic from DTOs and controller
