+++ src/main/kotlin/com/tarifit/auth/controller/AuthController.kt
@@ -1,7 +1,5 @@
 package com.tarifit.auth.controller
 
-import com.tarifit.auth.dto.AuthResponse
-import com.tarifit.auth.dto.ErrorResponse
 import com.tarifit.auth.dto.LoginRequest
 import com.tarifit.auth.dto.RegisterRequest
 import com.tarifit.auth.service.AuthService
@@ -16,53 +14,16 @@ class AuthController(private val authService: AuthService) {
     
     @PostMapping("/register")
     fun register(@Valid @RequestBody request: RegisterRequest): ResponseEntity<Any> {
-        return try {
-            val response = authService.register(request)
-            ResponseEntity.ok(response)
-        } catch (e: RuntimeException) {
-            ResponseEntity.badRequest().body(
-                ErrorResponse(
-                    error = "REGISTRATION_FAILED",
-                    message = e.message ?: "Registration failed"
-                )
-            )
-        }
+        return authService.handleRegister(request)
     }
     
     @PostMapping("/login")
     fun login(@Valid @RequestBody request: LoginRequest): ResponseEntity<Any> {
-        return try {
-            val response = authService.login(request)
-            ResponseEntity.ok(response)
-        } catch (e: RuntimeException) {
-            ResponseEntity.badRequest().body(
-                ErrorResponse(
-                    error = "LOGIN_FAILED",
-                    message = e.message ?: "Login failed"
-                )
-            )
-        }
+        return authService.handleLogin(request)
     }
     
     @GetMapping("/validate")
     fun validate(@RequestHeader("Authorization") authorization: String): ResponseEntity<Map<String, Any>> {
-        return try {
-            val token = authorization.removePrefix("Bearer ")
-            val isValid = authService.validateToken(token)
-            
-            ResponseEntity.ok(
-                mapOf(
-                    "valid" to isValid,
-                    "message" to if (isValid) "Token is valid" else "Token is invalid"
-                )
-            )
-        } catch (e: Exception) {
-            ResponseEntity.badRequest().body(
-                mapOf(
-                    "valid" to false,
-                    "message" to "Token validation failed"
-                )
-            )
-        }
+        return authService.handleValidate(authorization)
     }
 }
