+++ src/main/kotlin/com/tarifit/auth/dto/ErrorResponse.kt
@@ -3,5 +3,5 @@ package com.tarifit.auth.dto
 data class ErrorResponse(
     val error: String,
     val message: String,
-    val timestamp: String = java.time.Instant.now().toString()
+    val timestamp: String
 )
